// frontend/src/types/generated/models/Writing.ts
/**
 * Kotobalize API
 * API for the Kotobalize technical interview preparation app.
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class Writing {
    'id': number;
    'userId': number;
    'themeId': number;
    'content': string;
    'durationSeconds': number;
    'aiScore'?: number;
    'aiFeedbackOverall'?: string;
    'aiFeedbackClarity'?: string;
    'aiFeedbackAccuracy'?: string;
    'aiFeedbackCompleteness'?: string;
    'aiFeedbackStructure'?: string;
    'aiFeedbackConciseness'?: string;
    'createdAt': Date;
    'updatedAt': Date;

    static readonly discriminator: string | undefined = undefined;

    static readonly mapping: {[index: string]: string} | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "id",
            "baseName": "id",
            "type": "number",
            "format": "int64"
        },
        {
            "name": "userId",
            "baseName": "userId",
            "type": "number",
            "format": "int64"
        },
        {
            "name": "themeId",
            "baseName": "themeId",
            "type": "number",
            "format": "int64"
        },
        {
            "name": "content",
            "baseName": "content",
            "type": "string",
            "format": ""
        },
        {
            "name": "durationSeconds",
            "baseName": "durationSeconds",
            "type": "number",
            "format": "int32"
        },
        {
            "name": "aiScore",
            "baseName": "aiScore",
            "type": "number",
            "format": "int32"
        },
        {
            "name": "aiFeedbackOverall",
            "baseName": "aiFeedbackOverall",
            "type": "string",
            "format": ""
        },
        {
            "name": "aiFeedbackClarity",
            "baseName": "aiFeedbackClarity",
            "type": "string",
            "format": ""
        },
        {
            "name": "aiFeedbackAccuracy",
            "baseName": "aiFeedbackAccuracy",
            "type": "string",
            "format": ""
        },
        {
            "name": "aiFeedbackCompleteness",
            "baseName": "aiFeedbackCompleteness",
            "type": "string",
            "format": ""
        },
        {
            "name": "aiFeedbackStructure",
            "baseName": "aiFeedbackStructure",
            "type": "string",
            "format": ""
        },
        {
            "name": "aiFeedbackConciseness",
            "baseName": "aiFeedbackConciseness",
            "type": "string",
            "format": ""
        },
        {
            "name": "createdAt",
            "baseName": "createdAt",
            "type": "Date",
            "format": "date-time"
        },
        {
            "name": "updatedAt",
            "baseName": "updatedAt",
            "type": "Date",
            "format": "date-time"
        }    ];

    static getAttributeTypeMap() {
        return Writing.attributeTypeMap;
    }

    public constructor() {
    }
}